fix: Python: Semantic Kernel sample bug fix & updates (#579)
# Description

The Semantic Kernel Python Agent sample has a bug related to checking
for a non-existent `_thread_id`. The underlying SDK was fixed and the
sample does not have the change to properly check for the `.id` property
on the thread.

Additionally, we fixed the `on_intermediate_message` callback as part of
the Semantic Kernel agent framework, which properly emits the
`FunctionCallContent` and `FunctionResultContent` during function
calling. The sample is now updated to reflect this, along with updated
interim messsaging to the host during function call processing.

The PR also includes a few smaller updates in the README to make sure
the dev starts the host with the proper port that the SK agent server
uses.

Thank you for opening a Pull Request!
Before submitting your PR, there are a few things you can do to make
sure it goes smoothly:

- [X] Follow the [`CONTRIBUTING`
Guide](https://github.com/google/A2A/blob/main/CONTRIBUTING.md).
- [X] Make your Pull Request title in the
<https://www.conventionalcommits.org/> specification.
- [X] Ensure the tests and linter pass (Run `nox -s format` from the
repository root to format)
- [X] Appropriate docs were updated (if necessary)

Fixes #<issue_number_goes_here> 🦕

---------

Co-authored-by: Holt Skinner <holtskinner@google.com>