feat!: Add mTLS to SecuritySchemes, add oauth2 metadata url field, allow Skills to specify Security (#901)
# Description

This PR makes changes to our security representations in AgentCards:

1. Adds a new MutualTLSSecurityScheme type for indicating mTLS
connection requirements. This matches the same field in the OpenAPI
specification v3.1. This type is pretty lacking: there is no way to
indicate any requirements on the certificate that is presented by the
client. I expect we'll want to extend this in the future (for example,
we may want to allow an agent to indicate that a SPIFFE certificate must
be presented). For now, users are advised to put requirements in the
description field.
2. Adds a field for specifying a URL for OAuth2 provider metadata. This
is planned for OpenAPI Specification v3.2, which has not yet been
released. This allows clients to fetch additional information about the
OAuth provider, such as to discover if they support Dynamic Client
Registration.
3. Allows AgentSkills to specify a `security` field. This allows agents
to indicate security requirements for leveraging a particular skill.

I also clarified the usage and expectations for the `security` field in
the base AgentCard.

This change is marked backwards incompatible due to the addition of the
MutualTLSSecurityScheme, which adds a new type to an exhaustive enum
(the security scheme `anyOf`).

Release-As: 0.3.0

---------

Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>
Co-authored-by: Holt Skinner <holtskinner@google.com>